Refactor mortgage TUI: Separate app logic and UI components
- Moved application logic from main.rs to app.rs for better organization.
- Created a new ui.rs file to handle all UI rendering and interactions.
- Updated the App struct to manage state and logic separately from UI.
- Implemented CSV export functionality within the UI module.
- Enhanced field management and input handling in the App struct.
- Improved overall code structure and readability by separating concerns.
