+++ mortgage_core/Cargo.toml
@@ -5,8 +5,6 @@ edition = "2024"
 
 [dependencies]
 chrono = { workspace = true }
-rust_decimal = { workspace = true }
-rust_decimal_macros = { workspace = true }
 serde = { workspace = true }
 thiserror = { workspace = true }
 log = { workspace = true }
