+++ mortgage_core/src/euribor.rs
@@ -1,11 +1,10 @@
+use crate::error::MortgageError;
 use crate::models::EuriborTenor;
 use chrono::NaiveDate;
 
 /// Fetches the latest Euribor rate from ECB/Frankfurt for the given tenor.
 /// Returns the annual rate as a percentage (e.g., 4.5 for 4.5%).
-pub fn fetch_euribor(tenor: EuriborTenor) -> Result<f64, String> {
-    // ECB SDMX API endpoint for Euribor.
-    // Series key format: M.EXT_EURIBOR_6M (for 6-month Euribor)
+pub fn fetch_euribor(tenor: EuriborTenor) -> Result<f64, MortgageError> {
     let series = format!("M.EXT_EURIBOR_{}", tenor.as_str().to_uppercase());
     let url = format!(
         "https://sdw-wsrest.ecb.europa.eu/service/data/BSI/{}?lastNObservations=1&format=json",
@@ -15,16 +14,15 @@ pub fn fetch_euribor(tenor: EuriborTenor) -> Result<f64, String> {
     let resp = ureq::get(&url)
         .header("Accept", "application/json")
         .call()
-        .map_err(|e| format!("HTTP error: {}", e))?;
+        .map_err(|e| MortgageError::EuriborFetchError(format!("HTTP error: {}", e)))?;
 
     let body_text = resp
         .into_body()
         .read_to_string()
-        .map_err(|e| format!("Read body error: {}", e))?;
+        .map_err(|e| MortgageError::EuriborFetchError(format!("Read body error: {}", e)))?;
     let json: serde_json::Value = serde_json::from_str(&body_text)
-        .map_err(|e| format!("JSON parse error: {}", e))?;
+        .map_err(|e| MortgageError::EuriborFetchError(format!("JSON parse error: {}", e)))?;
 
-    // Navigate ECB SDMX JSON structure.
     let value = json
         .get("dataSets")
         .and_then(|v| v.get(0))
@@ -34,7 +32,7 @@ pub fn fetch_euribor(tenor: EuriborTenor) -> Result<f64, String> {
         .and_then(|v| v.get("0"))
         .and_then(|v| v.get(0))
         .and_then(|v| v.as_f64())
-        .ok_or("Could not extract Euribor value from response")?;
+        .ok_or_else(|| MortgageError::EuriborFetchError("Could not extract Euribor value from response".to_string()))?;
 
     Ok(value)
 }
@@ -53,7 +51,7 @@ impl EuriborCache {
     }
 
     /// Get rate, fetching if necessary. Caches for the current calendar day.
-    pub fn get_or_fetch(&mut self, tenor: EuriborTenor) -> Result<f64, String> {
+    pub fn get_or_fetch(&mut self, tenor: EuriborTenor) -> Result<f64, MortgageError> {
         let today = chrono::Local::now().date_naive();
         if let (Some(cached_tenor), Some(cached_date), Some(cached_rate)) =
             (self.tenor, self.fetched_at, self.rate)
